Split RHCOS into layers
This enhancement describes improvements to the way RHEL CoreOS (RHCOS)
is built so that it will better align with image mode for RHEL, all
while also providing benefits on the OpenShift side. Currently, RHCOS
is built as a single layer that includes both RHEL and OCP content. This
enhancement proposes splitting it into three layers. Going from bottom
to top:
1. the (RHEL-versioned) bootc layer (i.e. the base rhel-bootc image
   shared with image mode for RHEL)
2. the (RHEL-versioned) CoreOS layer (i.e. coreos-installer, ignition,
   afterburn, scripts, etc...)
3. the (OCP-versioned) node layer (i.e. kubelet, cri-o, etc...)

The terms "bootc layer", "CoreOS layer", and "node layer" will be used
throughout this enhancement to refer to these.

The details of this enhancement focus on doing the first split: creating
the node layer as distinct from the CoreOS layer (which will not yet be
rebased on top of a bootc layer). The two changes involved which most
affect OCP are:
1. bootimages will no longer contain OCP components (e.g. kubelet,
   cri-o, etc...)
2. the `rhel-coreos` payload image will be built in Prow/Konflux (as
   any other)

Tracked at: https://issues.redhat.com/browse/OCPSTRAT-1190