Fix "previously allocated" table
Whoever added this thought it would be clever to not bother aligning
the columns, and then promptly proceeded to make it clear why the
columns needed to be aligned by accidentally making the body have
more columns than the headings.

Remove the "Protocol" column since it's not really useful for ports
that aren't actually still used, and change "Since" to "Added" and
"Removed".