OKD: add snippet regarding OKD branding
The branding for OKD is different from Openshift and is controlled in
some places by go files which have the `//go:build !ocp` directive, but
the dockerfile hardcodes ocp. To get around this, the TAGS variable for
introduced in some repos to build the OKD variant of code. Document for
posterity.