udn, virt: clarify user stories / goals / non-goals
We realized we do not understand the use case for allowing the user to
specify the IP address for their VMs. Thus, we have postponed that goal
to a future release, while re-scoping the goal for this one to be about
persisting the IP address of the VM migrating from another
virtualization platform into OpenShift Virtualization.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>