Explicitly configure address management with UDNs
The current proposal for UDN's IPAM configuration is limited and would
not allow a simple expansion covering other use-cases. This patch
aims to make it more future-proof.

There are two limitations that this should solve:

1) The current proposal does not leave space for default subnets value
2) The current proposal does not leave space for granular control of IPAM

This patch should solve both by introducing a new field, that controls
how much of the IP configuration will be managed by OVN-K and how much is
left to the user.

Apart from introducing this new field, this patch also changes the
behavior of nil subnets. While with the current enhancement not setting
subnets means disabling IPAM, with this patch, not setting subnets is
not allowed when IP management is enabled.
This will allow us to introduce default subnet value in the future,
if we decide to go that path.

Co-authored-by: Petr Horacek <phoracek@redhat.com>
Signed-off-by: Nadia Pinaeva <n.m.pinaeva@gmail.com>