user-defined-network-segmentation: add VRF field to CUDN
Currently VRF names are generated in a way that can not be easily
identified and correlated to the UDNs.

Cluster admins may want predictable VFR names:
* VFR-Lite day 0 configuration
* Friendlier VRF matching of non auto non default target VRFs on route
  advertisements
* Consistent VRF name configuration around the fabric
* Same ease of use and level of customization as the default VRF

Provide that through an optional field of the CUDN. If not specified,
use the CUDN name if possible. Otherwise, use the internal genrated
names we have been using up until now.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>