OCPBUGS-1460: Update 2NO EP with relevant structure and content from OLA EP.
In this commit, I ran through all of the content in the OLA EP
(https://github.com/openshift/enhancements/pull/1674/files) and
brought over any details or structure that I found relevant or helpful
for the 2-node proposal. I also tweaked some sections with formatting
updates and updated some of the overall content to reflect the current
plan and understanding.