status: discuss web console changes and API absence
Web Console has resource discovery so we do not have to ship a plugin to
accomodate presence/absence of the API, it can handle it. Therefore we
can drop this from drawbacks (and keep just a brief mention about clients
having to gracefully degrade in API absence), and discuss the necessary
changed directly in the proposal (new Update UX in Web Console section).