CONVENTIONS: Explain no plan for node-role.kubernetes.io/control-plane taint
Adding a new node-role.kubernetes.io/control-plane label was
straightforward, because existing consumers could continue using the
legacy node-role.kubernetes.io/master label.  Removing the
node-role.kubernetes.io/master label would risk disrupting consumers.

But adding a new taint is disruptive, because folks who want to run on
control-plane nodes would need to grow tolerations for the incoming
taint.  Currently, a handful of components are tolerating the
control-plane taint [1], preparing for a world where that taint might
exist [2].  But the cost of dancing existing components (both
OpenShift and customer workloads) that would need to transition over
to the new taint is high enough that we don't expect the new taint to
ever exist.

[1]: https://github.com/openshift/cluster-kube-apiserver-operator/pull/1664
[2]: http://github.com/openshift/enhancements/pull/1583