gateway-api: Update to reflect Gateway API v1.2.1
Update goals and non-goals to say that we support Gateway API
v1.2.1 (rather than v0.5.1), that we support stable features (not
v1beta1), and that we do support ReferenceGrant (which is now stable
albeit beta).

Update the question about v1beta1 CRDs to add some historical context
and to note that the GA feature supports all v1beta1 and v1 APIs in
Gateway API v1.2.1.

This commit resolves NE-2041.

https://issues.redhat.com/browse/NE-2041

* enhancements/ingress/gateway-api-with-cluster-ingress-operator.md:
Update.