VEP 1827- Metrics Collection Sidecars for KubeVirt Components
This VEP outlines a proposale to implementing
dedicated metrics collection sidecars for
KubeVirt components to improve observability,
reduce resource overhead on core components,
and enable more flexible metrics collection and export
strategies.

This will also include moving out the recording rules
and alerts to a dedicated repository.

Created with the help of Cursor.
Signed-off-by: Shirly Radco <sradco@redhat.com>