AGENT-1253: Interactive installation using the agent-based
installer for disconnected environments

This enhancement proposes how to simplify cluster installations in
a disconnected environment using the agent-based, assisted-installer,
and appliance technologies.

This proposal is required for adding a new image to the OpenShift
release image. The new image contains the assisted-installer UI
for the new interactive UI driven flow.