Document CO's Conditions in one place
There were two sections about `co.status.condtions` when it was
introduced [1]. The [enhancements#1842](https://github.com/openshift/enhancements/pull/1842/files)
starts to use the GoDoc as one source of truth for those conditions
but it updated one section and left the other unchanged which is
removed by this change. If it leads to missing any essential point
(which I do not see it is the case), we should add it back into
the API docs and then sync it here.

In addition, we copied the doc for the two optional conditions
to keep the documentation complete.

[1]. https://github.com/hongkailiu/enhancements/commit/f8a9ee5510c8e93781df479f38416ea9761919d6#diff-3df7d241231880b4c05bfc1cced8f4b603409472b2cf35ce78bbe04e2fad302e