Add nodeservice-etc-hosts-injection
An extension to the DNS configuration that defines a list of services
and namespaces which are appended to the existing cluster image registry
entry stored in the SERVICES environment variable for the dns-resolver
daemonset running in the openshift-dns namespace. This allows node
services to resolve cluster services and connect to them. Potential use
cases are for augmenting monitoring leveraging a cluster hosted service
for any other means within a node scoped process.